added folder navigation and creation as well as updated deletion for both files and folders

+++ app/protected/fileUpload/page.tsx
@@ -2,8 +2,7 @@
 import { createClient } from '@/lib/supabase/client';
 import { User } from "@supabase/supabase-js";
 import {useEffect, useState} from "react";
-import {FaTrash} from "react-icons/fa";
-import { type FileObject } from "@supabase/storage-js"; // part of supabase-js
+import {FaTrash, FaFolder} from "react-icons/fa";
 
 // example using Supabase Auth Helpers for Next.js
 
@@ -27,6 +26,8 @@ export default function Page() {
     const supabase = createClient();
     const [currentFolderPath, setCurrentFolderPath] = useState<string>("root");
     const [currentFolderID, setCurrentFolderID] = useState<string>("");
+    const [previousPath, setPreviousPath] = useState<string>("root");
+    const [previousID, setPreviousID] = useState<string>("");
 
     const [user, setUser] = useState<User | null>(null);
     const [userItems, setUserItems] = useState<Item[]>([])
@@ -93,15 +94,34 @@ export default function Page() {
     };
 
 
-    async function removeFile(filename: string){
+    async function removeFile(id: string, filepath: string | null){
         if (!supabase || !user) {
             console.error('Supabase client or user not available yet')
             return
         }
-        const res = await fetch(`/protected/api/deleteFile?name=${encodeURIComponent(filename)}`, {
+        if(!filepath){
+            console.error("invalid filepath recieved for remove operation: " + filepath)
+            return
+        }
+        const res = await fetch(`/protected/api/deleteFile?id=${encodeURIComponent(id)}&filepath=${encodeURIComponent(filepath)}`, {
             method: "delete",
         });
-        fetchFiles(user);
+        fetchFiles(user, currentFolderID);
+    }
+
+    async function removeFolder(id: string, filepath: string | null){
+        if (!supabase || !user) {
+            console.error('Supabase client or user not available yet')
+            return
+        }
+        if(!filepath){
+            console.error("invalid filepath recieved for remove operation: " + filepath)
+            return
+        }
+        const res = await fetch(`/protected/api/deleteFolder?id=${encodeURIComponent(id)}&filepath=${encodeURIComponent(filepath)}`, {
+            method: "delete",
+        });
+        fetchFiles(user, currentFolderID);
     }
 
 
@@ -127,11 +147,62 @@ export default function Page() {
                 .from('user-files')
                 .upload(`${filepath}/${filename}`, file)
             if (error) throw error
-            fetchFiles(user);
+            fetchFiles(user, currentFolderID);
         }
 
         return;
     }
+
+    async function createFolder(folderName : string){
+        const isfolder = true;
+        const filepath = currentFolderPath;
+        const parentID = currentFolderID;
+        console.log({ parentid : parentID, desiredname : folderName, isfolder : isfolder, storagepath : filepath})
+        if (!supabase || !user) {
+            console.error('Supabase client or user not available yet')
+            return
+        }
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        const {data: queryData, error: queryError} = await supabase.rpc('upload_file',{ parentid : parentID, desiredname : folderName, isfolder : isfolder, mimetype : null, sizebytes : null, storagepath : filepath})
+        if (queryError) {
+            console.error("folder creation failed:", queryError.message);
+        } else {
+            console.log("folder creation succeeded!");
+            fetchFiles(user, currentFolderID);
+        }
+
+        return;
+    }
+
+     function enterFolder(fileID: string, filepath: string | null, filename: string){
+        if (!supabase || !user) {
+            console.error('Supabase client or user not available yet')
+            return
+        }
+        if (!filepath){
+            console.error("folder entry failure: filepath is null")
+            return
+        }
+        setPreviousPath(filepath);
+        setPreviousID(currentFolderID);
+
+        setCurrentFolderPath(filepath + '/' + filename);
+        setCurrentFolderID(fileID);
+        fetchFiles(user, fileID);
+    }
+
+    function navigateBack(){
+        if (!supabase || !user) {
+            console.error('Supabase client or user not available yet')
+            return
+        }
+
+        setCurrentFolderPath(previousPath)
+        setCurrentFolderID(previousID)
+        fetchFiles(user, previousID)
+    }
+
+
 
     async function downloadFile(fileID: string, filePath: string, filename : string){
         if (!supabase || !user) {
@@ -161,6 +232,17 @@ export default function Page() {
     return (
         <form>
             <p>Current Folder path: {currentFolderPath}</p>
+            <label
+                htmlFor="navigate-back"
+                onClick={() => navigateBack()}
+                className="inline-flex cursor-pointer items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0 bg-primary text-primary-foreground shadow hover:bg-primary/90 h-9 px-4 py-2">Back
+            </label>
+            <label
+                htmlFor="create-folder"
+                onClick={() => createFolder('newfolder')}
+                className="inline-flex cursor-pointer items-center justify-center gap-2 whitespace-nowrap rounded-md text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50 [&_svg]:pointer-events-none [&_svg]:size-4 [&_svg]:shrink-0 bg-primary text-primary-foreground shadow hover:bg-primary/90 h-9 px-4 py-2">New
+                Folder
+            </label>
             <div>
                 <input
                     id="file-upload"
@@ -183,21 +265,22 @@ export default function Page() {
             <div>
                 <h2>Files uploaded:</h2>
                 <ul>
-                     {userItems?.map((item) => (
+                    {userItems?.map((item) => (
                         <li key={item.id} className="flex items-center justify-between gap-2">
     <span
         className="cursor-pointer text-blue-600 hover:underline"
-        onClick={() => downloadFile(item.id, currentFolderPath, item.name)}
+        onClick={() => item.is_folder ? enterFolder(item.id, item.storage_path, item.name) : downloadFile(item.id, currentFolderPath, item.name)}
     >
       {item.name}
     </span>
                             <button
                                 type="button"
                                 className="p-1 text-red-500 hover:text-red-700"
-                                onClick={() => removeFile(item.name)}
+                                onClick={() => item.is_folder ? removeFolder(item.id, item.storage_path) : removeFile(item.id, item.storage_path)}
                             >
                                 <FaTrash size={16}/>
                             </button>
+                            {item.is_folder && (<FaFolder size={16}/>)}
                         </li>
                     ))}
                 </ul>
